+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 0.0.1
+## 0.0.2
 
-* TODO: Describe initial release.
+added options_page.dart <br>
+added various tests <br>
+(hopefully) fixed WebViewController issue
+
+## 0.0.3
+
+moved options_page.dart to proper directory (oops)
+
+## 0.0.4
+
+Many bugfixes, added ZugApp and Options/Lobby/Splash pages